Fix AF Control Tower missing filters via hooks page_js
Serve the dashboard from a cache-bustable public JS asset registered in
page_js, mount Company/Branch/Cost Center/Profit Center/UNLOCO/Fiscal Year
with make_control in a visible Parameters card, and stub tabPage.script so
stale inline page JS cannot hide filters.
