+++ logistics/air_freight/page/air_freight_control_tower/air_freight_control_tower.json
@@ -4,7 +4,7 @@
  "creation": "2026-07-23 00:00:00.000000",
  "docstatus": 0,
  "idx": 0,
- "modified": "2026-07-24 03:25:00.000000",
+ "modified": "2026-07-24 03:30:00.000000",
  "modified_by": "Administrator",
  "module": "Air Freight",
  "name": "air-freight-control-tower",
