+++ logistics/hooks.py
@@ -83,6 +83,7 @@ app_include_js = [
 # include js in page
 page_js = {
 	"workflow-center": "public/js/workflow_center.js",
+	"air-freight-control-tower": "public/js/air_freight_control_tower_page.js",
 }
 
 # include js in doctype views
